requirements.txt
absl-py==1.4.0
accelerate==0.31.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
analytics-python==1.4.post1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==1.10.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blendmodes==2024.1.1
-e git+https://github.com/z-x-yang/videogpt.git@33a81efb074893c01c188f836ac2ad082c6bb284#egg=boxmot&subdirectory=yolo_tracking-master
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.7
cmake==3.27.2
colorama==0.4.6
contourpy==1.1.0
crcmod==1.7
cryptography==42.0.7
cycler==0.11.0
dataclasses-json==0.5.14
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
diffusers==0.23.0
distlib==0.3.8
dnspython==2.6.1
easydict==1.10
einops==0.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
executing==1.2.0
facexlib==0.3.0
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.14.0
filterpy==1.4.5
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.2.0
future==1.0.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-vision==3.4.4
googleapis-common-protos==1.60.0
gradio==3.37.0
gradio_client==0.6.1
greenlet==3.0.3
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@edf3ed0de347e3922c2fedad8bd3d2ba7861a514#egg=groundingdino&subdirectory=GroundingDINO
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
identify==2.5.36
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
ipdb==0.13.13
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonmerge==1.9.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.0
langchain==0.0.261
langchain-experimental==0.0.17
langsmith==0.0.92
lapx==0.5.9
lark==1.1.9
lazy_loader==0.3
lightning-utilities==0.9.0
linkify-it-py==2.0.3
lit==16.0.6
llvmlite==0.42.0
loguru==0.7.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.3
mmcls==0.25.0
mmcv==2.1.0
mmcv-full==1.5.0
mmengine==0.10.4
mmsegmentation==0.27.0
model-index==0.1.11
monotonic==1.6
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.9.0
numba==0.59.1
numexpr==2.9.0
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
openai==0.27.8
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
orjson==3.10.3
oss2==2.17.0
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
Pillow==10.0.0
platformdirs==3.11.0
pre-commit==3.7.1
prettytable==3.9.0
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.9
pydantic_core==2.18.4
pydub==0.25.1
Pygments==2.16.1
PyNaCl==1.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.0.7
pytorch-ranger==0.1.1
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
resize-right==0.0.2
rich==13.4.2
rpds-py==0.18.1
rsa==4.9
ruff==0.4.7
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.4.1.post1
scipy==1.11.2
seaborn==0.13.2
-e git+https://github.com/IDEA-Research/Grounded-Segment-Anything.git@edf3ed0de347e3922c2fedad8bd3d2ba7861a514#egg=segment_anything&subdirectory=segment_anything
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sk-video==1.1.10
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.2
starlette==0.37.2
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2023.8.12
tiktoken==0.3.1
timm==0.4.12
tinycudann==1.7
tokenizers==0.13.3
tomesd==0.1.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.0+cu118
torch-optimizer==0.3.0
torchaudio==2.1.0+cu118
torchdiffeq==0.2.4
torchmetrics==1.1.0
torchsde==0.2.6
torchvision==0.16.0+cu118
tqdm==4.65.2
traitlets==5.9.0
trampoline==0.1.2
transformers==4.26.1
triton==2.1.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
ultralytics==8.0.124
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.3.7
whisper==1.1.10
yacs==0.1.8
yapf==0.40.2
yarl==1.9.2
zipp==3.16.2
